# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PropFlow API (auth-server + crm-api)",
    "version": "1.0.0",
    "contact": {
      "name": "PropFlow",
      "email": "bionex.product@gmail.com"
    },
    "license": {
      "name": "Proprietary — hiring assignment",
      "url": "https://github.com/your-github/propflow"
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local (npm run dev:auth / docker compose)"
    },
    {
      "url": "https://propflow-auth.up.railway.app",
      "description": "Railway (replace with your deployed URL)"
    },
    {
      "url": "http://localhost:4001",
      "description": "Local (npm run dev:crm / docker compose; via nginx on :8080)"
    },
    {
      "url": "https://propflow-crm.up.railway.app",
      "description": "Railway (replace with your deployed URL)"
    }
  ],
...
  "paths": {
...
    "/auth/invite": {
      "post": {
        "tags": [
          "Invites"
        ],
        "summary": "Invite a user into the tenant (ADMIN)",
        "description": "Creates a signed, single-use invite link valid for 24 h. No email is sent: the link is returned and logged. A newer invite for the same email revokes older pending ones. Alias of `POST /invites`.",
        "operationId": "createInvite",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "ADMIN"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invite created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "A user with this email exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "EMAIL_TAKEN",
                    "message": "A user with this email already exists",
                    "details": {
                      "fields": {
                        "email": "A user with this email already exists"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users of the tenant (ADMIN)",
        "operationId": "listUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "ADMIN"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TenantRole"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Matches name or email"
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PublicUser"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
    "/users/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "User id",
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get one user of the tenant (ADMIN)",
        "operationId": "getUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "ADMIN"
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUser"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Change role, rename, deactivate / reactivate (ADMIN)",
        "description": "The last active ADMIN of a tenant can be neither demoted nor deactivated (422 `LAST_ADMIN`; admin rows are locked FOR UPDATE so concurrent demotions cannot both pass). Deactivation or a role change revokes the user’s refresh tokens immediately.",
        "operationId": "updateUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "ADMIN"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUser"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "description": "Last admin guard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "LAST_ADMIN",
                    "message": "The last active admin cannot be demoted or deactivated",
                    "details": {
                      "fields": {
                        "role": "At least one active admin is required"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
    "/invites": {
      "get": {
        "tags": [
          "Invites"
        ],
        "summary": "List invites of the tenant, newest first (ADMIN)",
        "operationId": "listInvites",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "ADMIN"
        ],
        "responses": {
          "200": {
            "description": "Invites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Invite"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "post": {
        "tags": [
          "Invites"
        ],
        "summary": "Create an invite (ADMIN) — same as POST /auth/invite",
        "operationId": "createInvite2",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "ADMIN"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invite created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "Email taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
    "/invites/{id}/link": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Invite id",
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "get": {
        "tags": [
          "Invites"
        ],
        "summary": "Re-issue the signed link of a pending invite (\"Copy link\")",
        "operationId": "inviteLink",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "ADMIN"
        ],
        "responses": {
          "200": {
            "description": "Link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "description": "Invite not pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
    "/invites/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Invite id",
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "delete": {
        "tags": [
          "Invites"
        ],
        "summary": "Revoke a pending invite (ADMIN)",
        "operationId": "revokeInvite",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "ADMIN"
        ],
        "responses": {
          "200": {
            "description": "Revoked invite",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/X-Request-Id"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ]
      },
      "servers": [
        {
          "url": "http://localhost:4000",
          "description": "Local (npm run dev:auth / docker compose)"
        },
        {
          "url": "https://propflow-auth.up.railway.app",
          "description": "Railway (replace with your deployed URL)"
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT (RS256)",
        "description": "Access token from `/auth/login`, `/auth/refresh`, `/auth/register-tenant` or `/auth/accept-invite`. RS256, header carries `kid`; claims `sub`, `tid`, `role`, `name`, `jti`, `iat`, `exp`, `iss=propflow-auth`, `aud=propflow`. 60 s lifetime in the deployed build. Keep it in memory only."
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "pf_rt",
        "description": "Opaque 256-bit rotating refresh token (7 days). httpOnly; Secure; SameSite=None (web and API are on different sites); Path=/auth. Stored in Redis only as SHA-256. Never sent in a JSON body."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The one error shape returned by both services for every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "UNAUTHENTICATED",
                  "TOKEN_EXPIRED",
                  "INVALID_TOKEN",
                  "INVALID_CREDENTIALS",
                  "REFRESH_INVALID",
                  "REFRESH_REUSED",
                  "FORBIDDEN_ROLE",
                  "FORBIDDEN_ORIGIN",
                  "NOT_FOUND",
                  "VERSION_CONFLICT",
                  "DUPLICATE_LISTING",
                  "CONFLICT",
                  "EMAIL_TAKEN",
                  "SLUG_TAKEN",
                  "BUSINESS_RULE",
                  "LAST_ADMIN",
                  "IN_USE",
                  "TERMINAL_STATUS",
                  "INVITE_INVALID",
                  "ACCOUNT_DISABLED",
                  "RATE_LIMITED",
                  "INTERNAL",
                  "SERVICE_UNAVAILABLE"
                ],
                "description": "Stable machine-readable code"
              },
              "message": {
                "type": "string",
                "description": "Human-readable message (safe to show in the UI)"
              },
              "details": {
                "description": "Optional, code-specific. For VALIDATION_ERROR it is `ValidationErrorDetails`; for VERSION_CONFLICT it carries the current server copy.",
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ValidationErrorDetails"
                  },
                  {
                    "type": "object",
                    "additionalProperties": true
                  }
                ]
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        }
      },
      "ValidationErrorDetails": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Field path → first error message. The web app maps these onto react-hook-form inputs.",
            "example": {
              "priceInr": "Price must be greater than 0",
              "ownerPhone": "Enter a valid 10-digit Indian mobile number"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        },
        "required": [
          "fields"
        ]
      },
      "Role": {
        "type": "string",
        "enum": [
          "SUPER_ADMIN",
          "ADMIN",
          "MANAGER",
          "AGENT"
        ]
      },
      "TenantRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "MANAGER",
          "AGENT"
        ]
      },
...
      "PublicUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 3
          },
          "tenantId": {
            "type": "integer",
            "example": 1,
            "description": "null for SUPER_ADMIN",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "isActive": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "email",
          "name",
          "role",
          "isActive"
        ],
        "example": {
          "id": 3,
          "tenantId": 1,
          "email": "manager@skyline.dev",
          "name": "Neha Kapoor",
          "role": "MANAGER",
          "isActive": true,
          "lastLoginAt": "2026-10-01T04:31:10.000Z",
          "createdAt": "2026-09-01T10:00:00.000Z"
        }
      },
...
      "InviteRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "role": {
            "$ref": "#/components/schemas/TenantRole"
          }
        },
        "required": [
          "email",
          "name",
          "role"
        ],
        "example": {
          "email": "rahul@skyline.dev",
          "name": "Rahul Verma",
          "role": "AGENT"
        }
      },
      "Invite": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "tenantId": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/TenantRole"
          },
          "invitedBy": {
            "type": "integer",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "acceptedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "state": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REVOKED",
              "EXPIRED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InviteResponse": {
        "type": "object",
        "properties": {
          "invite": {
            "$ref": "#/components/schemas/Invite"
          },
          "inviteUrl": {
            "type": "string",
            "format": "uri",
            "description": "Signed, single-use, 24 h. Also written to the logs (no real email is sent)."
          },
          "inviteToken": {
            "type": "string",
            "description": "RS256 JWT (typ=invite, aud=propflow:invite, jti=invite id)"
          }
        },
        "required": [
          "invite",
          "inviteUrl",
          "inviteToken"
        ],
        "example": {
          "invite": {
            "id": 12,
            "tenantId": 1,
            "email": "rahul@skyline.dev",
            "name": "Rahul Verma",
            "role": "AGENT",
            "invitedBy": 2,
            "expiresAt": "2026-10-02T04:31:10.000Z",
            "acceptedAt": null,
            "revokedAt": null,
            "state": "PENDING",
            "createdAt": "2026-10-01T04:31:10.000Z"
          },
          "inviteUrl": "https://propflow-yourname.web.app/accept-invite?token=eyJhbGciOiJSUzI1NiIs…",
          "inviteToken": "eyJhbGciOiJSUzI1NiIs…"
        }
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/TenantRole"
          },
          "isActive": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          }
        },
        "minProperties": 1,
        "example": {
          "role": "MANAGER"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "400 — request failed validation against the shared Zod schema",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Request validation failed",
                "details": {
                  "fields": {
                    "email": "Invalid email"
                  },
                  "issues": [
                    {
                      "path": "email",
                      "code": "invalid_string",
                      "message": "Invalid email"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "401 — missing, invalid or expired access token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "missing": {
                "value": {
                  "error": {
                    "code": "UNAUTHENTICATED",
                    "message": "Authentication required"
                  }
                }
              },
              "expired": {
                "summary": "Access token expired → client calls /auth/refresh and retries",
                "value": {
                  "error": {
                    "code": "TOKEN_EXPIRED",
                    "message": "Access token expired"
                  }
                }
              },
              "invalid": {
                "value": {
                  "error": {
                    "code": "INVALID_TOKEN",
                    "message": "Invalid access token"
                  }
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "403 — authenticated, but the role may not do this",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "FORBIDDEN_ROLE",
                "message": "Your role does not allow this action"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "404 — does not exist, belongs to another tenant, or (for an AGENT) is not assigned to you. These cases are deliberately indistinguishable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "NOT_FOUND",
                "message": "Property not found"
              }
            }
          }
        }
      },
      "Unprocessable": {
        "description": "422 — a business rule was violated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "BUSINESS_RULE",
                "message": "Some listings were not found; nothing was changed",
                "details": {
                  "missingIds": [
                    999999
                  ]
                }
              }
            }
          }
        }
      },
...
    },
    "parameters": {
      "XRequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "description": "Optional correlation id (8–64 chars `[A-Za-z0-9._:-]`). Generated when absent and echoed back in the response header; appears on every log line.",
        "schema": {
          "type": "string",
          "example": "web-7f3a9c21"
        }
      }
    },
    "headers": {
      "X-Request-Id": {
        "description": "Correlation id for this request (search logs with it)",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    }
  },
...
}